backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
pillow>=10.2.0
python-multipart>=0.0.9
numpy>=1.26.0
